test: ignore postcss-parser-tests until after 7.x update
