+++ package.json
@@ -50,11 +50,11 @@
     "eslint-config-shellscape": "^1.0.0",
     "eslint-plugin-filenames": "^1.2.0",
     "is-absolute-url": "^2.1.0",
-    "less": "^2.6.1",
+    "less": "^3.8.1",
     "lint-staged": "^7.2.2",
     "node-fetch": "^2.2.0",
     "nyc": "^13.0.1",
-    "postcss-parser-tests": "^5.0.11",
+    "postcss-parser-tests": "^6.3.0",
     "pre-commit": "^1.2.2",
     "standard-version": "^4.4.0",
     "url-join": "^4.0.0"
